projects/deepseek/.deepseek/requirements.txt
"fastapi[standard]"
ollama
python-jose
python-dotenv
psycopg2-binary
sqlmodel